+++ Editor/MarkdownViewer/Scripts/Layout/IActions.cs
@@ -1,0 +1,15 @@
+﻿////////////////////////////////////////////////////////////////////////////////
+
+using UnityEngine;
+
+
+namespace MG.MDV
+{
+    public interface IActions
+    {
+        Texture FetchImage(string url);
+
+
+        void SelectPage(string url);
+    }
+}