+++ Editor/MarkdownViewer/Scripts/MarkdownAsset.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+
+namespace MG.MDV
+{
+    public class MarkdownAsset : TextAsset
+    {
+    }
+}